requirements.txt
streamlit==1.53.1
google-generativeai==0.8.6
streamlit-ace-0.1.1